requirements.txt
attrs==18.2.0
Automat==0.7.0
boto3==1.9.23
botocore==1.12.23
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.6.1
coloredlogs==10.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
dill==0.2.8.2
docutils==0.14
#en-core-web-md==2.0.0
future==0.16.0
gevent==1.3.7
greenlet==0.4.15
humanfriendly==4.16.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
klein==17.10.0
matplotlib==2.2.3
msgpack==0.5.6
msgpack-numpy==0.4.4.1
murmurhash==0.28.0
numpy==1.15.2
packaging==18.0
pathlib==1.0.1
Pillow==5.3.0
plac==0.9.6
preshed==1.0.1
PyHamcrest==1.9.0
pyparsing==2.2.2
pytesseract==0.2.5
python-crfsuite==0.9.6
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
rasa-nlu==0.13.7
regex==2017.4.5
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
simplejson==3.16.0
six==1.11.0
sklearn-crfsuite==0.3.6
spacy==2.0.12
tabulate==0.8.2
thinc==6.10.3
toolz==0.9.0
tqdm==4.26.0
Twisted==18.7.0
typing==3.6.6
ujson==1.35
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.5.0
